% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cluster_similarity}
\alias{cluster_similarity}
\title{Cluster similarity matrix}
\usage{
cluster_similarity(x, expr, exprTest, distanceMethod,
  clustMethod = c("hclust", "protoclust"), cutMethod = c("dynamic", "gap",
  "fixed"), nClusters, method = c("complete", "average", "ward.D2", "single",
  "ward.D", "mcquitty", "median", "centroid"), K.max = 10, B = 50, nPC)
}
\arguments{
\item{x}{similarity matrix. must have non-NULL dimnames i.e., the rows and
columns should be labelled, e.g. "Gene1, Gene2, ..."}

\item{expr}{gene expression data (training set). rows are people, columns are
genes}

\item{exprTest}{gene expression test set}

\item{distanceMethod}{one of "euclidean","maximum","manhattan", "canberra",
"binary","minkowski" to be passed to \code{\link[stats]{dist}} function. If
missing, then this function will take 1-x as the dissimilarity measure.
This functionality is for diffCorr,diffTOM, fisherScore matrices which need
to be converted to a distance type matrix.}

\item{clustMethod}{Cluster the data using hierarchical clustering or
prototype clustering. Defaults \code{clustMethod="hclust"}. Other option is
\code{\link[protoclust]{protoclust}}, however this package must be
installed before proceeding with this option}

\item{cutMethod}{what method to use to cut the dendrogram. \code{'dynamic'}
refers to \code{\link[dynamicTreeCut]{}} library. \code{'gap'} is
Tibshirani's gap statistic \code{\link[cluster]{clusGap}} using the
\code{'Tibs2001SEmax'} rule. \code{'fixed'} is a fixed number specified by the
\code{nClusters} argument}

\item{nClusters}{number of clusters. Only used if \code{cutMethod = fixed}}

\item{method}{the agglomeration method to be used. This should be (an
unambiguous abbreviation of) one of "ward.D", "ward.D2", "single",
"complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC)
or "centroid" (= UPGMC).}

\item{K.max}{the maximum number of clusters to consider, must be at least
two. Only used if \code{cutMethod='gap'}}

\item{B}{integer, number of Monte Carlo (“bootstrap”) samples. Only used if \code{cutMethod='gap'}}
}
\description{
Return cluster membership of each predictor. This function is
  called internally by the \code{\link{generate_data}} and
  \code{\link{generate_data_mars}} function
}

