% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{clusterSimilarity}
\alias{clusterSimilarity}
\title{Cluster similarity matrix and return cluster membership of each gene
as well as average expression per module, 1st PC of each module, and 1st PC
of each module from Test dataset}
\usage{
clusterSimilarity(x, expr, exprTest, distanceMethod, clustMethod = c("hclust",
  "protoclust"), cutMethod = c("dynamic", "gap", "fixed"), nClusters,
  method = c("complete", "average", "ward.D2", "single", "ward.D", "mcquitty",
  "median", "centroid"), K.max = 10, B = 50, nPC)
}
\arguments{
\item{x}{similarity matrix. must have non-NULL dimnames i.e., the rows and
columns should be labelled preferable "Gene1, Gene2, ..."}

\item{expr}{gene expression data (training set). rows are people, columns are
genes}

\item{exprTest}{gene expression test set}

\item{distanceMethod}{one of "euclidean","maximum","manhattan", "canberra",
"binary","minkowski" to be passed to \code{dist} function. If missing, then
this function will take 1-x as the dissimilarity measure. This
functionality is for diffCorr,diffTOM, fisherScore matrices which need to be
converted to a distance type matrix.}

\item{cutMethod}{what method to use to cut the dendrogram. \code{dynamic}
refers to dynamicTreeCut library, \code{gap} is Tibshirani's gap statistic
\code{fixed} is a fixed number specified by the \code{nClusters} argument}

\item{nClusters}{number of clusters. Only used if \code{cutMethod = fixed}}

\item{method}{the agglomeration method to be used. This should be (an
unambiguous abbreviation of) one of "ward.D", "ward.D2", "single",
"complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC)
or "centroid" (= UPGMC).}

\item{clusterMethod}{how to cluster the data}
}
\description{
Cluster similarity matrix and return cluster membership of each gene
as well as average expression per module, 1st PC of each module, and 1st PC
of each module from Test dataset
}

