% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclust-package.R
\docType{package}
\name{eclust-package}
\alias{eclust-package}
\alias{_PACKAGE}
\alias{{eclust}-package}
\title{eclust: Time Series Clustering using Energy Distance}
\description{
This package implements the multivariate time series clustering algorithm from Davis et al. 2022 (doi: 10.1111/jtsa.12688). The method is based on a pairwise multivariate distance covaraince metric followed by a hierarchical clustering procedure to obtain the final clusters.
}
\references{
Richard A. Davis, Leon Fernandes, Konstantinos Fokianos
"Clustering multivariate time series using energy distance",
Journal of Time Series Analysis, 44: 487-504.
https://doi.org/10.1111/jtsa.12688
}
\author{
\strong{Maintainer}: Leon Fernandes \email{leon.fernandes@columbia.com} (\href{https://orcid.org/0000-0002-7192-9939}{ORCID})

Authors:
\itemize{
  \item Richard A. Davis \email{rdavis@stat.columbia.edu} (\href{https://orcid.org/0000-0001-6936-3772}{ORCID})
  \item Konstantinos Fokianos \email{fokianos.konstantinos@ucy.ac.cy} (\href{https://orcid.org/0000-0002-0051-711X}{ORCID})
}

}
\keyword{internal}
