% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{edist.matrix}
\alias{edist.matrix}
\title{Lagged Energy Distance for Matrices}
\usage{
\method{edist}{matrix}(x, sizes, group_ids = NULL, type, lag, weight, a)
}
\arguments{
\item{x}{a \code{matrix}}

\item{sizes}{numeric vector of sample sizes}

\item{group_ids}{character vector of names for each grouped time series}

\item{type}{character determining the type of lagged energy distance.}

\item{lag}{number of lags.}

\item{a}{index for energy distance. Must be in \eqn{(0, 2]}.}
}
\description{
See \link{edist}.
}
