% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extractPC}
\alias{extractPC}
\title{This is a modified version of firstPC which was actually giving the first 2 PCs
with no other option. This function is more flexible and the nPC argument is used.
currently only nPC = 1 and nPC = 2 are supported}
\usage{
extractPC(x_train, colors, x_test, y_train, y_test, impute = TRUE, nPC,
  excludeGrey = FALSE, grey = if (is.numeric(colors)) 0 else "grey",
  subHubs = TRUE, trapErrors = FALSE, returnValidOnly = trapErrors,
  softPower = 6, scale = TRUE, verbose = 0, indent = 0)
}
\description{
This is a modified version of firstPC which was actually giving the first 2 PCs
with no other option. This function is more flexible and the nPC argument is used.
currently only nPC = 1 and nPC = 2 are supported
}

