% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{GDP Data of 23 Countries}
\format{
\code{gdp}
A \link[tsibble:tsibble]{tsibble::tsibble} of 920 rows and 3 columns:
\describe{
\item{country}{Country name}
\item{year}{Year}
\item{gdp}{GDP value}
}
}
\source{
\url{https://www.conference-board.org/us/}
}
\usage{
gdp
}
\description{
A subset of the Conference Board containing Real GDP of 23 most developed countries.
}
\keyword{datasets}
