% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.similarity}
\alias{plot.similarity}
\title{Function to generate heatmap}
\usage{
\method{plot}{similarity}(x, color = viridis(100), truemodule, active, ...)
}
\description{
x matrix of true correlation (P x P matrix where P is the number
  of genes). Must be object of class similarity
}

