% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{s_mars_reg}
\alias{s_mars_reg}
\title{Title}
\usage{
s_mars_reg(x_train, x_test, y_train, y_test, s0, model = c("MARS"),
  expFamily = c("gaussian", "binomial"), true_beta, topgenes = NULL,
  stability = F, filter = F, include_E = F, include_interaction = F,
  p = 1000, filter_var = F, ...)
}
\arguments{
\item{x_train}{\code{ntrain x p} matrix of simulated training set where
\code{ntrain} is the number of training observations  and \code{p} is total
number of predictors. This matrix needs to have named columns representing
the feature names or the gene names}

\item{x_test}{\code{ntest x p} matrix of simulated training set where
\code{ntest} is the number of training observations  and \code{p} is total
number of predictors. This matrix needs to have named columns representing
the feature names or the gene names}

\item{y_train}{numeric vector of length \code{ntrain} representing the
responses for the training subjects. If continuous then you must set
\code{exp_family = "gaussion"}. For \code{exp_family="binomial"} should be
either a factor with two levels, or a two-column matrix of counts or
proportions (the second column is treated as the target class; for a
factor, the last level in alphabetical order is the target class)}

\item{y_test}{numeric vector of length \code{ntest} representing the
responses for the test subjects. If continuous then you must set
\code{exp_family = "gaussion"}. For \code{exp_family="binomial"} should be
either a factor with two levels, or a two-column matrix of counts or
proportions (the second column is treated as the target class; for a
factor, the last level in alphabetical order is the target class).}

\item{s0}{chracter vector of the active feature names, i.e., the features in
\code{x_train} that are truly associated with the response.}

\item{model}{Regression model to be fit on cluster summaries. Default is
\code{model="lasso"} which corresponds to glmnet mixing parameter
\code{alpha=1}. \code{model="elasticnet"} corresponds to glmnet mixing
parameter \code{alpha=0.5}, \code{model="mcp"} and \code{model="scad"} are
the non-convex models from the \code{\link[ncvreg]{}} package}

\item{topgenes}{List of features to keep if \code{filter=TRUE}. Default is
\code{topgenes = NULL} which means all features are kept for the analysis}

\item{stability}{Should stability measures be calculated. Default is
\code{stability=FALSE}. See details}

\item{filter}{Should analysis be run on a subset of features. Default is
\code{filter = FALSE}}

\item{include_E}{Should the environment variable be included in the
regression analysis. Default is \code{include_E = TRUE}}

\item{include_interaction}{Should interaction effects between the features in
\code{x_train} and the environment variable be fit. Default is
\code{include_interaction=TRUE}}
}
\description{
Title
}

