% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{s_response}
\alias{s_response}
\title{Generate True Response vector for Linear Simulation}
\usage{
s_response(n, n0, p, genes, E, signal_to_noise_ratio = 1,
  include_interaction = F, beta = NULL)
}
\arguments{
\item{n}{Total number of subjects}

\item{n0}{Total number of unexposed subjects}

\item{p}{Total number of genes (or covariates)}

\item{genes}{nxp matrix of the genes or covariates}

\item{E}{binary 0,1, vector of the exposure/environment variable}

\item{signal_to_noise_ratio}{a numeric variable for the signal to noise ratio}

\item{include_interaction}{Logical. Should the response include the
interaction between E and the genes (for the non-zero \code{beta}
coefficient vector)}

\item{beta}{true beta coefficient vector. Assumes this vector is in the same
order as the \code{genes}.}
}
\value{
a data.frame/data.table containing the response and the design
  matrix. Also an object of class \code{expression}
}
\description{
Given the true beta vector, covariates and environment variable this function
generates the linear response with specified signal to noise ratio.
}

