% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{s_response_mars}
\alias{s_response_mars}
\title{Generate True Response vector for Non-Linear Simulation}
\usage{
s_response_mars(n, n0, p, genes, E, signal_to_noise_ratio = 1, truemodule,
  nActive)
}
\arguments{
\item{n}{total number of subjects}

\item{n0}{total number of subjects with E=0}

\item{p}{number of genes in design matrix}

\item{signal_to_noise_ratio}{signal to noise ratio, default is 1}

\item{truemodule}{numeric vector of the true module membership used in the
\code{s_response_mars} function. Modules 3 and 4 are active in the
response. See \code{s_response_mars} function for details.}

\item{nActive}{number of active genes in the response used in the
\code{s_response_mars}}
}
\value{
a data.frame/data.table containing the response and the design
  matrix. Also an object of class \code{expression}
}
\description{
Given the covariates and environment variable this function
generates the nonlinear response with specified signal to noise ratio.
}
\note{
See Bhatnagar et al (2017+) for details on how the response is simulated.
}

