% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{u_extract_selected_earth}
\alias{u_extract_selected_earth}
\title{Get selected terms from an earth object}
\usage{
u_extract_selected_earth(obj)
}
\arguments{
\item{obj}{object of class \code{earth} returned by the
\code{\link[earth]{earth}} function}
}
\value{
character vector of selected terms from the MARS model
}
\description{
function to extract the selected terms from an earth object
}
\details{
called internally by the \code{\link{s_mars_separate}} and
  \code{\link{s_mars_clust}} functions
}

