% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{u_fisherZ}
\alias{fisherTransform}
\alias{u_fisherZ}
\title{Calculate Fisher's Z Transformation for Correlations}
\usage{
u_fisherZ(n0, cor0, n1, cor1)

fisherTransform(n1, r1, n2, r2)
}
\arguments{
\item{n0}{number of unexposed subjects}

\item{cor0}{correlation matrix of unexposed covariate values. Should be
dimension pxp}

\item{n1}{number of exposed subjects}

\item{cor1}{correlation matrix of exposed covariate values. Should be
dimension pxp}

\item{r1}{correlation for unexposed}

\item{r2}{correlation for exposed}
}
\description{
Calculate Fisher's Z transformation for correlations. This can
  be used as an alternative measure of similarity. Used in the
  \code{s_generate_data} function
}
\note{
\code{fisherTransform} is called internally by \code{u_fisherZ} function
}

