% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{edist}
\alias{edist}
\title{Energy Distance}
\usage{
edist(.data, type = c("upto", "paired"), lag = 0L, a = 1, weights = NULL)
}
\arguments{
\item{.data}{\code{tsibble} object.}

\item{type}{character representing type of lagged energy distance.}

\item{lag}{number of lags.}

\item{a}{coefficient for energy distance.}

\item{weights}{optional vector of weights to use when \code{type} is "paired".}
}
\value{
a \code{dist} object
}
\description{
Energy Distance
}
