% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{edist.tbl_ts}
\alias{edist.tbl_ts}
\title{Lagged Energy Distance for Tsibbles}
\usage{
\method{edist}{tbl_ts}(x, type, lag, weights, a)
}
\arguments{
\item{x}{a \code{tsibble} object.}

\item{type}{character determining the type of lagged energy distance.}

\item{lag}{number of lags.}

\item{weights}{optional vector of length \code{lag} + 1 when \code{type} is 2.
Overrides \code{lag} when incompatible.}

\item{a}{index for energy distance. Must be in \eqn{(0, 2]}.}
}
\description{
See \link{edist}.
}
